add plot
